+++ Automate/TrackedItemCollection.cs
@@ -2,10 +2,10 @@ using System.Collections.Generic;
 using System.Linq;
 using StardewValley;
 
-namespace Pathoschild.Stardew.Automate.Framework
+namespace Pathoschild.Stardew.Automate
 {
     /// <summary>An item stack which wraps an underlying collection of stacks.</summary>
-    internal class TrackedItemCollection : ITrackedStack
+    public class TrackedItemCollection : ITrackedStack
     {
         /*********
         ** Properties