+++ Automate/Framework/Machines/Objects/SlimeIncubatorMachine.cs
@@ -1,3 +1,4 @@
+using Microsoft.Xna.Framework;
 using StardewValley;
 using SObject = StardewValley.Object;
 
@@ -19,8 +20,9 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         /// <summary>Construct an instance.</summary>
         /// <param name="machine">The underlying machine.</param>
         /// <param name="location">The location containing the machine.</param>
-        public SlimeIncubatorMachine(SObject machine, GameLocation location)
-            : base(machine, location)
+        /// <param name="tile">The tile covered by the machine.</param>
+        public SlimeIncubatorMachine(SObject machine, GameLocation location, Vector2 tile)
+            : base(machine, location, tile)
         {
             int minutesUntilReady = Game1.player.professions.Contains(2) ? 2000 : 4000;
             this.Recipes = new IRecipe[] {