+++ Automate/release-notes.md
@@ -1,10 +1,10 @@
 [← back to readme](README.md)
 
 # Release notes
-## Upcoming release
-* Extended Automate API to simplify custom machines.
+## 1.11.1
+* Improved Automate API to simplify custom machines.
 * Fixed error when the game has invalid fish data.
-* Fixed tree tappers having the wrong automation position in Automate 1.11.
+* Fixed tree tapper automation issue in Automate 1.11.
 * Fixed gates not working as connectors.
 
 ## 1.11